Finish fixing issues with config file relative path location and finish remained of reorganization
