+++ README.md
@@ -22,17 +22,19 @@ slack: _Yeah, welcome to the club, pal_
     - Under "Interactive Messages", paste the HTTPS_FORWARDING_URL, and append "/interactive" to the url.
     - Under "Slash Commands", add "/pollease". Use HTTPS_FORWARDING_URL, and append "/pollease" to the url.
 5. Configure pollease, the app server
-    - Under "Basic Information", copy Client ID and Client Secret to config_local.py
+    - Under "Basic Information", copy Client ID and Client Secret to config_local.ini
 6. Run `sh build.sh`
     - This compiles the docker image
 7. Run `sh run.sh`
     - This starts a new container with the pollease server image
-8. Add pollease to your Slack team
+8. Alternatively run `sh rerun.sh` to do steps 6 and 7 in one command.
+9. Add pollease to your Slack team
     - Edit add-to-slack.html and insert your Client ID
     - Open add-to-slack.html in a browser and click the button. 
-9. If everything goes correctly, the app should now be added to your team and the /pollease command should be functional. All HTTP requests will now be tunnelled to your localhost.
+10. If everything goes correctly, the app should now be added to your team and the /pollease command should be functional. All HTTP requests will now be tunnelled to your localhost.
 
 ## Tests
 1. To run tests, you must have already built the pollease image via step 6 of the setup process.
 2. Run `sh build_test.sh` to build the test image.
 3. Run `sh test.sh` to run the test package in a container.
+4. Alternatively run `sh retest.sh` to do steps 2 and 3 in one command.
