+++ app/domain/custom_exceptions.py
