+++ app/resources/command_router.py
@@ -1,7 +1,7 @@
 """Returns the expected function based on the command text."""
 
-from .custom_exceptions import CommandParsingException
-from .constants import VALID_COMMANDS, COMMAND_CLOSE, COMMAND_CREATE, ERR_PARSING_COMMAND
+from app.domain.custom_exceptions import CommandParsingException
+from app.domain.constants import VALID_COMMANDS, COMMAND_CLOSE, COMMAND_CREATE, ERR_PARSING_COMMAND
 
 from .command_parser import __deserialize_text
 from .pollease_commands import create_poll, close_poll
