+++ tests/Dockerfile
@@ -4,6 +4,6 @@ RUN [ "pip", "install", "unittest2" ]
 
 COPY . /usr/src/tests/
 
-WORKDIR /usr/src
+WORKDIR "/usr/src"
 
 CMD [ "python", "-m", "unittest2", "discover" ]
