+++ tests/domain/__init__.py
@@ -1,0 +1,1 @@
+#Initializes this directory
